+++ anyblok/model/event.py
@@ -6,6 +6,7 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 from .plugins import ModelPluginBase
+from anyblok.mapper import ModelAttribute
 
 
 class EventPlugin(ModelPluginBase):
@@ -43,3 +44,26 @@ class EventPlugin(ModelPluginBase):
             ev = events[model][event]
             if val not in ev:
                 ev.append(val)
+
+
+class SQLAlchemyEventPlugin(ModelPluginBase):
+
+    def transform_base_attribute(self, attr, method, namespace, base,
+                                 transformation_properties,
+                                 new_type_properties):
+        """declare in the registry the sqlalchemy event
+
+        :param attr: attribute name
+        :param method: method pointer of the attribute
+        :param namespace: the namespace of the model
+        :param base: One of the base of the model
+        :param transformation_properties: the properties of the model
+        :param new_type_properties: param to add in a new base if need
+        """
+        if not hasattr(method, 'is_an_sqlalchemy_event_listener'):
+            return
+        elif method.is_an_sqlalchemy_event_listener is True:
+            self.registry._sqlalchemy_known_events.append(
+                (method.sqlalchemy_listener,
+                 namespace,
+                 ModelAttribute(namespace, attr)))