+++ doc/CHANGES.rst
@@ -27,6 +27,7 @@ Futur
 * [REF]: **hybrid_method** become an ``anyblok.model.plugin``
 * [REF]: adapter of mapper_args and table_args become an ``anyblok.model.plugin``
 * [REF]: **event** become an ``anyblok.model.plugin``
+* [REF]: **sqlachemy event** become an ``anyblok.model.plugin``
 
 0.9.10 (2017-09-23)
 -------------------